[Issue #45] Preparatory end to end test as a safety net.
Before introducing any changes to the existing JLBH, the end to end test
has been created. It integrates JLBH and System.out as nothing can be changed
in the production code at this stage to be sure that the test shows the original behaviour
and that this expected behaviour stays unmodified when introducing the new feature.