[Issue #45] Allow passing non standard print stream to test the output.
System.out as a default option is a sane choice for JLBH as the most
frequent (only?) use case is displaying the JLBH result on the screen.
However, hardcoded System.out reference as the only print stream output in JLBH
may be inconvenient when the output can be ignored or parsed further.

By keeping the existing contructor and adding additional one that allows
a client to pass a custom implementation of PrintStream we add this flexibility
without affecting existing clients as behavior stays exactly the same.